One ft script. Add A100 results
